Adding subfolders and spliting eltwise.py into 3 subfolders.
